# [BACKLOG-22705][CLEANUP] GlobalMessageUtil optimizations and performance improvements