**Stats prior to this fix:**
`GlobalMessageUtil.getBundle() ....................... 473,847 ms`
`GlobalMessageUtil.calculateString() ................... 6,243 ms`
`GlobalMessageUtil.getActiveLocales() .................. 6,165 ms`
`GlobalMessageUtil.getLocaleString() ..................... 573 ms`
`GlobalMessageUtil.buildHashKey() ........................ 480 ms`
`GlobalMessageUtil.<clinit> .............................. 203 ms`
_Total build time: 10:38 min_

**Stats after the fix:**
`GlobalMessageUtil$GlobalMessageControl.newBundle() ...... 235 ms`
`GlobalMessageUtil.calculateString() ..................... 200 ms`
`GlobalMessageUtil.getActiveLocales() .................... 192 ms`
`GlobalMessageUtil.getBundle() ............................ 99 ms `
`GlobalMessageUtil.isMissingKey()	......................... 98 ms`
`GlobalMessageUtil.<clinit> ................................  0 ms`
_Total build time: 02:31 min_

@pentaho/r2d2 @mbatchelor @tmcsantos @emartin-pentaho