@pentaho-amessier 2018-05-10
@lgrill-pentaho it looks like more than it actually is, a good chunk of it is some old cleanup, and updated/re-formatted Javadoc. The most significant change was to move the bundle creation to the controller class, where we have some built in fallback for free, hence I was able to trim the active locale set. That’s the change that seemed to have the most significant performance impact. Second, I got rid of the “potential match” concept, which basically looked at all the locales and returned the best match for a given key, based on the order of active locales, rather than return the first match, which is quicker. Lastly, I replaced String.format, which is inefficient, with StringBuilder, this may not have had any visible impact, but definitely doesn’t hurt.