@wingman-pentaho 2018-05-10
### Build Completed

:fire: This pull request has some issues. It would be preferable to fix them in order for it to be just perfect. See below for more details. Some links are also available below for further assistance in addressing those issues.

#### Build Commands

```sh
mvn -B -fn -f 'pom.xml' -pl 'core' -P '!assemblies' -Dsurefire.runOrder=alphabetical -Daudit -amd clean install
```

#### Cleanup Commands
```sh
mvn -B -f 'pom.xml' -pl 'core' -P '!assemblies' -amd build-helper:remove-project-artifact
```

#### Changed files

```console
  core/src/main/java/org/pentaho/di/i18n/GlobalMessageUtil.java
  core/src/test/java/org/pentaho/di/i18n/GlobalMessageUtilTest.java
```

#### Unit Test Coverage

These statistics help you identify how your changes have affected the coverage of the following files. If a file is not in this list, then its coverage was not affected by your changes. To get some help interpreting these metrics, please refer to [Jacoco's documentation](http://eclemma.org/jacoco/trunk/doc/counters.html).

<details>
<summary>:warning: <b>Coverage Changes</b> (click to expand)</summary>

###### org.pentaho.di.i18n.GlobalMessageUtil
 * Instruction Change: **-1.5%** :small_red_triangle_down:
 * Line Change: **-5.4%** :small_red_triangle_down:

###### :new: org.pentaho.di.i18n.GlobalMessageUtil.GlobalMessageControl
 * Branch Coverage: **64.3%** :exclamation:
 * Complexity Coverage: **72.7%** :exclamation:
 * Line Coverage: **75%** :exclamation:

###### org.pentaho.di.core.injection.bean.BeanInjector
 * Branch Change: **-0.7%** :small_red_triangle_down:
 * Complexity Change: **-1.2%** :small_red_triangle_down:

###### org.pentaho.di.trans.steps.dimensionlookup.DimensionLookupMeta
 * Method Change: **-1%** :small_red_triangle_down:

###### org.pentaho.di.trans.steps.replacestring.ReplaceStringMeta
 * Instruction Change: **-0.1%** :small_red_triangle_down:

###### org.pentaho.di.trans.steps.setvalueconstant.SetValueConstantMeta
 * Instruction Change: **-0.2%** :small_red_triangle_down:

###### org.pentaho.di.trans.steps.sortedmerge.SortedMergeMeta
 * Instruction Change: **-0.2%** :small_red_triangle_down:

###### org.pentaho.di.trans.steps.uniquerows.UniqueRowsMeta
 * Instruction Change: **-0.1%** :small_red_triangle_down:

###### org.pentaho.di.trans.steps.webservices.WebServiceMeta
 * Branch Change: **-1.7%** :small_red_triangle_down:
 * Complexity Change: **-1.2%** :small_red_triangle_down:
 * Instruction Change: **-0.1%** :small_red_triangle_down:

###### org.pentaho.di.ui.repository.pur.repositoryexplorer.model.UIEERepositoryUser
 * Branch Change: **-8.3%** :small_red_triangle_down:
 * Complexity Change: **-6.7%** :small_red_triangle_down:
 * Instruction Change: **-0.9%** :small_red_triangle_down:
 * Line Change: **-3.1%** :small_red_triangle_down:

</details>