@101000 2021-05-05
There is also a major performance issue when using MessageFormat.format API in this class .. can you consider switching to StringBuilder ?